# [android] improve ToolbarExtensions.UpdateIconColor performance