--- a/src/Core/AndroidNative/maui/src/main/java/com/microsoft/maui/PlatformInterop.java
+++ b/src/Core/AndroidNative/maui/src/main/java/com/microsoft/maui/PlatformInterop.java
@@ -2,11 +2,14 @@ package com.microsoft.maui;
 
 import android.content.Context;
 import android.content.res.ColorStateList;
+import android.graphics.BlendMode;
+import android.graphics.BlendModeColorFilter;
 import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.Paint;
 import android.graphics.Path;
 import android.graphics.PathEffect;
+import android.graphics.PorterDuff;
 import android.graphics.Typeface;
 import android.graphics.drawable.Drawable;
 import android.graphics.drawable.PaintDrawable;
@@ -26,6 +29,7 @@ import android.widget.LinearLayout;
 
 import androidx.annotation.ColorInt;
 import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
 import androidx.appcompat.widget.SearchView;
 import androidx.appcompat.widget.TintTypedArray;
 import androidx.coordinatorlayout.widget.CoordinatorLayout;
@@ -222,6 +226,42 @@ public class PlatformInterop {
             .attach();
 
         return pager;
+    }
+
+    /**
+     * Call setColorFilter on a Drawable, passing in (int)Microsoft.Maui.FilterMode
+     * Calls the appropriate methods for Android API 29/Q+
+     * @param drawable android.graphics.Drawable
+     * @param color android.graphics.Color
+     * @param mode (int)Microsoft.Maui.FilterMode
+     */
+    public static void setColorFilter(@NonNull Drawable drawable, int color, int mode) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            drawable.setColorFilter(new BlendModeColorFilter(color, getBlendMode(mode)));
+        } else {
+            drawable.setColorFilter(color, getPorterMode(mode));
+        }
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.Q)
+    static BlendMode getBlendMode(int mode) {
+        // NOTE: keep in sync with src/Core/src/Primitives/FilterMode.cs
+        switch (mode) {
+            case 0: return BlendMode.SRC_IN;
+            case 1: return BlendMode.MULTIPLY;
+            case 2: return BlendMode.SRC_ATOP;
+            default: throw new RuntimeException("Invalid Mode");
+        }
+    }
+
+    static PorterDuff.Mode getPorterMode(int mode) {
+        // NOTE: keep in sync with src/Core/src/Primitives/FilterMode.cs
+        switch (mode) {
+            case 0: return PorterDuff.Mode.SRC_IN;
+            case 1: return PorterDuff.Mode.MULTIPLY;
+            case 2: return PorterDuff.Mode.SRC_ATOP;
+            default: throw new RuntimeException("Invalid Mode");
+        }
     }
 
     private static void prepare(RequestBuilder<Drawable> builder, Target<Drawable> target, Boolean cachingEnabled, ImageLoaderCallback callback) {