--- a/src/Core/src/Primitives/FilterMode.cs
+++ b/src/Core/src/Primitives/FilterMode.cs
@@ -1,9 +1,10 @@
 namespace Microsoft.Maui
 {
+	// Keep in sync with src/Core/AndroidNative/maui/src/main/java/com/microsoft/maui/PlatformInterop.java
 	public enum FilterMode
 	{
-		SrcIn,
-		Multiply,
-		SrcAtop
+		SrcIn = 0,
+		Multiply = 1,
+		SrcAtop = 2,
 	}
 }