@jonathanpeppers 2023-09-06
I added more icons to the gallery to double-check the overflow icon color. I think it looks the same as before:

![image](https://github.com/dotnet/maui/assets/840039/fee206b5-f38e-4ac0-9c62-1a37618bb762)